Debug Forward Unit
Discover and solve the bug about forward unit.
The data cannot be fowarded completely.
For example, when the previous data in WB are written back to register
file, and the forwarding occurs in the EX stage. It is unable to access
the data that just put into the register file, and cause data hazard.

Solution : Add one more pipeline stage, called BUF stage in my code,
whose pipeline registers are used to store the data of WB stage to
complete the forwarding.